+++ external_push/models.py
@@ -230,7 +230,7 @@ class GenericPushTarget(models.Model):
         # We've got the data (in a json'ed string) - lets send it
         return string_to_send
 
-    def send_data(self):
+    def send_data(self) -> bool:
         # self.data_to_push() returns a JSON-encoded string which we will push directly out
         json_data = self.data_to_push()
 
@@ -247,7 +247,16 @@ class GenericPushTarget(models.Model):
         else:
             raise ValueError("Invalid target type specified for push target")
 
-        return False  # Should never get here, but just in case something changes later
+    def check_target_host(self):
+        if len(self.target_host) > 8:
+            if self.target_host[:7] == "http://":
+                pass
+            elif self.target_host[:8] == "https://":
+                pass
+            else:
+                self.target_host = "http://" + self.target_host
+                self.save()
+
 
 
 class BrewersFriendPushTarget(models.Model):
@@ -557,8 +566,18 @@ class BrewfatherPushTarget(models.Model):
 
         r = requests.post(self.logging_url, data=json_data, headers=headers)
         return True  # TODO - Check if the post actually succeeded & react accordingly
-        
-        
+
+    def check_logging_url(self):
+        if len(self.logging_url) > 8:
+            if self.logging_url[:7] == "http://":
+                pass
+            elif self.logging_url[:8] == "https://":
+                pass
+            else:
+                self.logging_url = "http://" + self.logging_url
+                self.save()
+
+
 class ThingSpeakPushTarget(models.Model):
     class Meta:
         verbose_name = "ThingSpeak Push Target"