+++ external_push/tasks.py
@@ -3,6 +3,7 @@ from __future__ import absolute_import, unicode_literals
 from huey import crontab
 from huey.contrib.djhuey import periodic_task, task, db_periodic_task, db_task
 from external_push.models import GenericPushTarget, BrewersFriendPushTarget, BrewfatherPushTarget, ThingSpeakPushTarget, GrainfatherPushTarget
+from django.core.exceptions import ObjectDoesNotExist
 
 import datetime, pytz, time
 from django.utils import timezone
@@ -14,20 +15,20 @@ from requests.models import MissingSchema
 def generic_push_target_push(target_id):
     try:
         push_target = GenericPushTarget.objects.get(id=target_id)
-    except:
-        # TODO - Replace with ObjNotFound
+    except ObjectDoesNotExist:
         return None
 
-    push_target.send_data()
-
+    try:
+        push_target.send_data()
+    except MissingSchema:
+        push_target.check_target_host()
     return None
 
 @db_task()
 def brewers_friend_push_target_push(target_id):
     try:
         push_target = BrewersFriendPushTarget.objects.get(id=target_id)
-    except:
-        # TODO - Replace with ObjNotFound
+    except ObjectDoesNotExist:
         return None
 
     push_target.send_data()
@@ -38,11 +39,13 @@ def brewers_friend_push_target_push(target_id):
 def brewfather_push_target_push(target_id):
     try:
         push_target = BrewfatherPushTarget.objects.get(id=target_id)
-    except:
-        # TODO - Replace with ObjNotFound
+    except ObjectDoesNotExist:
         return None
 
-    push_target.send_data()
+    try:
+        push_target.send_data()
+    except MissingSchema:
+        push_target.check_logging_url()
 
     return None
 
@@ -50,8 +53,7 @@ def brewfather_push_target_push(target_id):
 def thingspeak_push_target_push(target_id):
     try:
         push_target = ThingSpeakPushTarget.objects.get(id=target_id)
-    except:
-        # TODO - Replace with ObjNotFound
+    except ObjectDoesNotExist:
         return None
 
     push_target.send_data()
@@ -62,8 +64,7 @@ def thingspeak_push_target_push(target_id):
 def grainfather_push_target_push(target_id):
     try:
         push_target = GrainfatherPushTarget.objects.get(id=target_id)
-    except:
-        # TODO - Replace with ObjNotFound
+    except ObjectDoesNotExist:
         return None
 
     try: